requirements.txt
pandas
numpy
sqlalchemy
psycopg2-binary  # For PostgreSQL
jupyter
python-dotenv
matplotlib
seaborn